add goalie role/tactic to basic def, make pretty-lines
